Release 3.3.1e (#1015)
** Bugfix 
* Bugfix : no filter on delete document
* Bugfix : change order to validate values
* Bugfix : DocIdRef was not returned
* Bugfix : add missing setter method for manual parameter
* Bugfix : change value formula to manage parentheses
* Bugfix : We don't send output for the API and Myddleware UI
* Bugfix : change the way to get the environment
* Bugfix : enhance function changeValue and changeMultiValue
* Bugfix : dateref value
* Bugfix : Manage space into formula changeValue and changeMultiValue
* Bugfix : Rename function checkParentDocument to checkParentDocuments for rule and job classes
* Bugfix : Add http-client library
* Bugfix : Myddleware 3.3 : Change dockerfile
* Bugfix : clearSendData (manage fields 'source_date_modified', 'id_doc_myddleware', 'Myddleware_element_id' )
* Bugfix : output management
* Bugfix : manage return null value for check methods
* Bugfix : change the type attribute from private to protected
* Bugfix : init $emailAddresses attribute

** Connector
* SugarCRM : deletion management
* SugarCRM : a duplicate field on email address module
* Moodle : imporve error management
* Moodle : manage create only fields
* Moodle : fix issue on new custom field function
* Salesforce : add rule mode Update only (U)
* Airtable : fix bug on reference date
* Airtable : fix ref date field name

+++ src/Solutions/solution.php
@@ -78,9 +78,9 @@ class solutioncore
     protected Connection $connection;
     protected ParameterBagInterface $parameterBagInterface;
     protected EntityManagerInterface $entityManager;
-    private DocumentRepository $documentRepository;
-    private RuleRelationShipRepository $ruleRelationshipsRepository;
-    private FormulaManager $formulaManager;
+    protected DocumentRepository $documentRepository;
+    protected RuleRelationShipRepository $ruleRelationshipsRepository;
+    protected FormulaManager $formulaManager;
 
     public function __construct(
         LoggerInterface $logger,
@@ -397,6 +397,7 @@ class solutioncore
     public function createData($param): array
     {
         try {
+			$result = array();
             // For every document
             foreach ($param['data'] as $idDoc => $record) {
                 try {
@@ -405,6 +406,10 @@ class solutioncore
 
                     // Check control before create
                     $record = $this->checkDataBeforeCreate($param, $record, $idDoc);
+					// No action if null is returned
+					if ($record === null) {
+						continue;
+					}
                     // Call create method
                     $recordId = $this->create($param, $record, $idDoc);
 
@@ -481,6 +486,7 @@ class solutioncore
     public function updateData($param): array
     {
         try {
+			$result = array();
             // For every document
             foreach ($param['data'] as $idDoc => $record) {
                 try {
@@ -492,6 +498,10 @@ class solutioncore
                     }
                     // Check control before create
                     $record = $this->checkDataBeforeUpdate($param, $record, $idDoc);
+					// No action if null is returned
+					if ($record === null) {
+						continue;
+					}
                     // Call create methode
                     $recordId = $this->update($param, $record, $idDoc);
 
@@ -535,6 +545,7 @@ class solutioncore
     public function deleteData($param): array
     {
         try {
+			$result = array();
             // For every document
             foreach ($param['data'] as $idDoc => $record) {
                 try {
@@ -543,6 +554,10 @@ class solutioncore
                     }
                     // Check control before delete
                     $record = $this->checkDataBeforeDelete($param, $record);
+					// No action if null is returned
+					if ($record === null) {
+						continue;
+					}
                     // Call delete methode
                     $recordId = $this->delete($param, $record);
